requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
requests>=2.32.0
geopandas>=1.0.0
shapely>=2.0.0
pyproj>=3.6.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
openpyxl>=3.1.0
